% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posall.R
\name{posall}
\alias{posall}
\title{Sample Size of AEBSD for Postive and Overall}
\usage{
posall(rtrue,raux,ppv,lam1e,lam0e,lam1c,lam0c,tau,raccrual,powerdesired,alpha,costs)
}
\arguments{
\item{rtrue}{the prevalence rates of patients with positive true biomarker}

\item{raux}{the prevalence rates of patients with positive auxiliary biomarker}

\item{lam1e}{hazrd rate for biomarker-positive subpopulation in the experimental treatment group}

\item{lam0e}{hazrd rate for biomarker-negative subpopulation in the experimental treatment group}

\item{lam1c}{hazrd rate for biomarker-positive subpopulation in the control treatment group}

\item{lam0c}{hazrd rate for biomarker-negative subpopulation in the control treatment group}

\item{tau}{minimum follow-up time}

\item{raccrual}{accrual rate}

\item{powerdesired}{desired design power}

\item{alpha}{type-I error (for each test)}

\item{costs}{cost vector including costscr (scrrening cost), costtrt (average treatment cost), costfol
(follow-up cost) and costtest (biomarker test cost) for each randomized patient}
}
\value{
(1) The probability of randomizing patients with positive and negative auxiliary variable;
(2) Number of randomized patients for AEBSD (3) Screening number for AEBSD
(4) Number of randomized patients for BSD (5) Patient ratio (6) Cost ratio (7) Screening ratio
}
\description{
Calculate minimum sample size needed of AEBSD for testing both treatment effect in biomarker positve and
overall treatment effect, and its comparison with BSD
}
